[WebConsoleV2] Extract types out form WorkflowJobFlowChart
