+++ web_console_v2/client/src/views/Workflows/CreateWorkflow/SteptTwoConfig/index.tsx
@@ -4,7 +4,7 @@ import { ReactFlowProvider, useStoreState, useStoreActions } from 'react-flow-re
 import { useToggle } from 'react-use';
 import JobFormDrawer, { JobFormDrawerExposedRef } from '../../JobFormDrawer';
 import WorkflowJobsFlowChart, { ChartExposedRef } from 'components/WorkflowJobsFlowChart';
-import { ChartNode, ChartNodes, JobNodeStatus } from 'components/WorkflowJobsFlowChart/helpers';
+import { ChartNode, ChartNodes, JobNodeStatus } from 'components/WorkflowJobsFlowChart/types';
 import GridRow from 'components/_base/GridRow';
 import { Button, message, Modal, Spin } from 'antd';
 import { Redirect, useHistory, useParams } from 'react-router-dom';
@@ -164,6 +164,14 @@ const CanvasAndForm: FC<WorkflowCreateProps> = ({ isInitiate, isAccept }) => {
 
     setConfigValue(nextValue);
   }
+  async function validateCurrentValues() {
+    if (!currNode) return;
+    const isValid = await drawerRef.current?.validateCurrentForm();
+    chartRef.current?.updateNodeStatusById({
+      id: currNode.id,
+      status: isValid ? JobNodeStatus.Success : JobNodeStatus.Warning,
+    });
+  }
   /** 🚀 Initiate create request */
   async function submitToCreate() {
     if (!checkIfAllJobConfigCompleted()) {
@@ -212,13 +220,8 @@ const CanvasAndForm: FC<WorkflowCreateProps> = ({ isInitiate, isAccept }) => {
     const prevNode = currNode;
     if (currNode && prevNode) {
       // Validate & Save current form before go another job
+      await validateCurrentValues();
       await saveCurrentValues();
-
-      const isValid = await drawerRef.current?.validateCurrentForm();
-      chartRef.current?.updateNodeStatusById({
-        id: currNode.id,
-        status: isValid ? JobNodeStatus.Success : JobNodeStatus.Warning,
-      });
     }
 
     // Turn target node status to configuring
@@ -232,21 +235,13 @@ const CanvasAndForm: FC<WorkflowCreateProps> = ({ isInitiate, isAccept }) => {
 
   // ---------- Handlers ----------------
   async function onCanvasClick() {
+    await validateCurrentValues();
     saveCurrentValues();
     toggleDrawerVisible(false);
-    const isValid = await drawerRef.current?.validateCurrentForm();
-    chartRef.current?.updateNodeStatusById({
-      id: currNode?.id!,
-      status: isValid ? JobNodeStatus.Success : JobNodeStatus.Warning,
-    });
   }
   async function onCloseDrawer() {
+    await validateCurrentValues();
     saveCurrentValues();
-    const isValid = await drawerRef.current?.validateCurrentForm();
-    chartRef.current?.updateNodeStatusById({
-      id: currNode?.id!,
-      status: isValid ? JobNodeStatus.Success : JobNodeStatus.Warning,
-    });
     setSelectedElements([]);
   }
   function onGoNextJob() {
