+++ web_console_v2/client/src/views/Workflows/WorkflowDetail/JobExecutionPODs.tsx
@@ -6,7 +6,7 @@ import i18n from 'i18n';
 import { Button } from 'antd';
 import StateIndicator, { StateTypes } from 'components/StateIndicator';
 import { useTranslation } from 'react-i18next';
-import { NodeDataRaw } from 'components/WorkflowJobsFlowChart/helpers';
+import { NodeDataRaw } from 'components/WorkflowJobsFlowChart/types';
 
 const Container = styled.div`
   margin-top: 30px;
